Factor out version code from build.py
